refactor: rename miniscript core to compiler
This renames the core miniscript pipeline to src/compiler, flattens correctness/malleability and aligns docs and imports with the new structure.
