--- src/miniscript/types/index.ts
@@ -1,5 +1,0 @@
-// Copyright (c) 2026 Jose-Luis Landabaso - https://bitcoinerlab.com
-// Distributed under the MIT software license
-
-export * from './correctness';
-export * from './malleability';
